🚚[Rename] rating에서 rating-display로 폴더명 수정 #38
