+++ src/components/rating-display/RatingDisplay.stories.tsx
