+++ src/components/rating-display/RatingDisplay.tsx
